+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## v21.10.0 alpha.4 (WIP)
 ### ⚠ BREAKING CHANGES
 * `Orion` was removed in favor of `EEA` as *PrivateTxManager* in chain APIs
+* Following ETH transaction properties types has been updated:
+    - `value` expects an HEX value prefix by "0x".
+    - `gasPrice` expects an HEX value prefix by "0x".
+* Following Faucet request params has been modified:
+    - `amount` expects an HEX value prefix by "0x".
+    - `maxBalance` expects an HEX value prefix by "0x".
 
 ## v21.10.0 alpha.3 (2021-11-24)
 ### 🆕 Features